Migrate to typescript and add basic test
