+++ .github/workflows/main.yml
@@ -1,0 +1,22 @@
+name: Test
+on:
+  - push
+
+jobs:
+  jest:
+    name: jest
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        node-version: [ 16 ]
+        cache: ['npm']
+
+    steps:
+      - uses: actions/checkout@master
+      - uses: actions/setup-node@v2
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - run: npm ci
+      - run: npm test
