--- fixtures/example.html
@@ -1,1 +1,0 @@
-<h1>Cool stuff</h1><div>cool</div>
