--- fixtures/sink.blade.php
@@ -1,8 +1,0 @@
-<h1>
-
-Here's some very poorly formatted
-    {{'blade'}} with weird {{
-        $echoParts + 1 / $insideOfIt}}
-</h1>
-
-<p class="hover:text-blue-500 py-2 text-blue-600 px-4 bg-blue-100">Hello, world!</p>
