+++ package.json
@@ -1,17 +1,25 @@
 {
     "name": "@ryangjchandler/prettier-plugin-blade",
     "devDependencies": {
-        "esbuild": "^0.14.14",
-        "tailwindcss": "^3.0.16"
+        "tailwindcss": "^3.0.16",
+        "@types/jest": "^27.0.0",
+        "@types/node": "^17.0.4",
+        "jest": "^27.0.1",
+        "ts-jest": "^27.0.2",
+        "ts-node": "^10.0.0",
+        "typescript": "^4.3.2"
     },
     "scripts": {
-        "build": "esbuild ./src/index.js --format=cjs --bundle --outfile=./dist/plugin.js --platform=node",
         "watch": "npm run build -- --watch",
-        "test": "esbuild ./test/index.js --bundle --outfile=./dist/test.js && node ./dist/test.js"
+        "test": "jest"
     },
     "dependencies": {
         "@prettier/plugin-php": "^0.17.6",
         "prettier": "^2.5.1",
         "prettier-plugin-tailwindcss": "^0.1.4"
+    },
+    "jest": {
+        "preset": "ts-jest",
+        "testRegex": ".test.ts$"
     }
 }
