+++ src/lang/index.ts
@@ -1,0 +1,12 @@
+import { TokenType, Token } from './token'
+import { Lexer } from './lexer'
+import { Parser } from './parser'
+import * as Nodes from './nodes'
+
+export default {
+    TokenType,
+    Token,
+    Lexer,
+    Parser,
+    Nodes
+}
