--- src/parser/index.js
@@ -1,12 +1,0 @@
-import { TokenType, Token } from './token'
-import { Lexer } from './lexer'
-import { Parser } from './parser'
-import * as Nodes from './nodes'
-
-export {
-    TokenType,
-    Token,
-    Lexer,
-    Parser,
-    Nodes
-}
