+++ src/types.ts
@@ -1,0 +1,1 @@
+import type * as Prettier from "prettier";
