+++ tests/echo/echo.blade.php
@@ -1,0 +1,2 @@
+    <h1>
+    Here's some very poorly formatted {{'html'}} with weird {{$echoParts + 1 / $insideOfIt}}</h1>
